+++ sandbox/src/platformer.ts
@@ -338,8 +338,8 @@ game.input.pointers.primary.on('move', (ev) => {
 
 game.input.pointers.primary.on('wheel', (ev) => {
   pointer.pos.setTo(ev.x, ev.y);
-  game.currentScene.camera.z += (ev.deltaY / 1000);
-  game.currentScene.camera.z = ex.Util.clamp(game.currentScene.camera.z, .05, 100);
+  game.currentScene.camera.zoom += (ev.deltaY / 1000);
+  game.currentScene.camera.zoom = ex.Util.clamp(game.currentScene.camera.zoom, .05, 100);
 })
 // Turn on debug diagnostics
 game.showDebug(false);
@@ -494,7 +494,7 @@ player.rotation = 0;
 
 // Health bar example
 var healthbar = new ex.Actor(0, -70, 140, 5, new ex.Color(0, 255, 0));
-player.add(healthbar);
+player.addChild(healthbar);
 // player.onPostDraw = (ctx: CanvasRenderingContext2D) => {
 //   ctx.fillStyle = 'red';
 //   ctx.fillRect(0, 0, 100, 100);